finfit_backend/requirements.txt
Django==5.2.4
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-nested-routers==0.95.0
openai
openpyxl==3.1.5
pandas==2.3.3
requests==2.32.5
